requirements.txt
APScheduler==3.11.0
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.2.3
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
greenlet==3.2.3
idna==3.10
iniconfig==2.1.0
packaging==25.0
pillow==11.2.1
playwright==1.52.0
pluggy==1.6.0
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.9.0
pytest==8.4.0
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
requests==2.32.4
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
stripe==12.2.0
text-unidecode==1.3
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
